# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Control Layer — API gatewaya
  version: 0.1.0
  description: |
    Kontrakt publiczny gatewaya (VISION.md §7). Dokument opisuje stan zaimplementowany w backendzie.

    **Zaimplementowane:** `POST /v1/chat/completions` (+ nagłówek `X-Session-Id`, zapisywany w audycie),
    `GET /actuator/health`, audyt `/api/audit/**` (lista, rekord, eksport CSV/JSON, weryfikacja łańcucha).
    `/api/dashboard`, `/api/auth/login|me|logout` (sesja przeglądarki, ciasteczko SESSION; HTTP Basic dalej działa dla maszyn).
    **Jeszcze nie:** reszta `/api/**` (dashboard, polityki, modele).

    **Uwierzytelnianie:** HTTP Basic na kontach lokalnych z rolami (docs/auth/). Konta demo są w
    `backend/config/users.yaml` i działają tylko w profilu `local`. Rola konta decyduje, które
    modele może wołać (`backend/config/policy.yaml`).

    **Kontrakt odpowiedzi:** każda odpowiedź `/v1/chat/completions` (także błędy 400, 403, 502) ma
    kształt `GuardedChatResponse`. Wyjątek: 401 z warstwy Spring Security może nie mieć treści.
servers:
  - url: http://localhost:8000
    description: Lokalnie (profil local, `./gradlew bootRun`)
  - url: http://raspberrypi.local:8000
    description: Raspberry Pi w sieci (plan wdrożenia)
security:
  - basicAuth: []
paths:
...
  /api/audit/events:
    get:
      summary: Lista rekordów audytu (najnowsze pierwsze)
      description: |
        Rola ADMIN. Jeden rekord na żądanie /v1/chat/completions (także allow) z pełną ścieżką
        kontroli; bez treści promptów i odpowiedzi. Paginacja keyset: `before` = `nextCursor`.
      operationId: auditEvents
      parameters:
        # Listy: parametr powtórzony (action=block&action=redact) = rekord ma którąkolwiek z wartości.
        - { name: action, in: query, schema: { type: array, items: { type: string } }, explode: true }
        - { name: principal, in: query, schema: { type: array, items: { type: string } }, explode: true }
        - { name: model, in: query, schema: { type: array, items: { type: string } }, explode: true }
        - { name: blockedBy, in: query, schema: { type: array, items: { type: string } }, explode: true }
        - { name: sessionId, in: query, description: 'Zawiera (bez wielkości liter; % i _ to zwykłe znaki).', schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: before, in: query, schema: { type: integer, format: int64 } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        '200':
          description: Strona rekordów.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/AuditEvent' } }
                  nextCursor: { type: integer, format: int64, nullable: true }
        '401': { description: Brak uwierzytelnienia. }
        '403': { description: Rola inna niż ADMIN. }
...
  /api/audit/events/{requestId}:
    get:
      summary: Jeden rekord audytu po requestId z odpowiedzi czatu
      operationId: auditEvent
      parameters:
        - { name: requestId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Rekord.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditEvent' }
        '404': { description: Brak rekordu. }
  /api/audit/export:
    get:
      summary: Eksport audytu (CSV albo JSON) z tymi samymi filtrami co lista
      description: Maks. 10 000 rekordów. Komórki CSV zaczynające się od = + - @ dostają prefiks ' (formula injection).
      operationId: auditExport
      parameters:
        - { name: format, in: query, schema: { type: string, enum: [csv, json], default: csv } }
      responses:
        '200': { description: Plik do pobrania (Content-Disposition attachment). }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Konto lokalne z rolą (app_user). Hasło jest sprawdzane przez BCrypt.
  schemas:
    AuditEvent:
      type: object
      description: Rekord audytu (AuditEventView.java). Bez treści wiadomości.
      properties:
        seq: { type: integer, format: int64 }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
        principal: { type: string, nullable: true }
        role: { type: string, nullable: true }
        sessionId: { type: string, nullable: true, description: Z nagłówka X-Session-Id, zsanitizowany. }
        model: { type: string, nullable: true }
        action: { type: string, example: redact }
        blockedBy: { type: string, nullable: true }
        httpStatus: { type: integer }
        latencyMs: { type: integer, format: int64 }
        usage: { $ref: '#/components/schemas/Usage' }
        messageCount: { type: integer }
        trace: { type: array, items: { $ref: '#/components/schemas/ControlTrace' } }
        recordHash: { type: string, description: HMAC-SHA256 ogniwa łańcucha. }
...
    Usage:
      type: object
      required: [promptTokens, completionTokens]
      properties:
        promptTokens:
          type: integer
          example: 18
        completionTokens:
          type: integer
          example: 40
    ControlTrace:
      type: object
      description: Wynik jednej kontroli w tym żądaniu (do X-ray w UI).
      required: [policy, kind, action, latencyMs]
      properties:
        policy:
          type: string
          description: |
            Nazwa polityki/kontroli, np. `model.allowlist`, `policy.model-access`, `upstream-error`,
            `request.validation` albo id guardu (`PII-001`).
          example: model.allowlist
        kind:
          type: string
          enum: [deterministic, semantic]
        action:
          $ref: '#/components/schemas/Action'
        latencyMs:
          type: integer
          format: int64
          example: 3
        detail:
          type: string
          nullable: true
          description: Opis dla człowieka (np. powód blokady). Brak danych wrażliwych.
    Action:
      type: string
      enum: [allow, redact, block]
...